Set import paths correctly
Old import paths were valid for pymice-analyzer itself but not the
resulting project. When calling load_data() after using import_all()
there is a "no attribute" error, so import it explicitly for now. Also,
the Docker Jupyter Notebook is having trouble parsing the symlink to
the data dir, so use its actual relative path for now.